Move backgrounds to experimental module
